fix(setup-gcloud): Only clean install backup dir if it exists

+++ setup-gcloud/src/setup-gcloud.js
@@ -71,8 +71,10 @@ const installComponents = async (toolPath) => {
   });
   // Ensure install backups are empty.
   const installBackup = path.join(toolPath, '.install', '.backup');
-  fs.rmSync(installBackup, { recursive: true });
-  fs.mkdirSync(installBackup, { recursive: true });
+  if (fs.existsSync(installBackup)) {
+    fs.rmSync(installBackup, { recursive: true });
+    fs.mkdirSync(installBackup, { recursive: true });
+  }
   glob.sync('**/__pycache__', {
     cwd: toolPath,
     dot: true,
